+++ packages/dumb/src/modules/plot-methods/charts/events/event-superposition.ts
@@ -28,16 +28,16 @@
 import _ from 'lodash'
 import * as d3 from 'd3'
 
-import { type Pool } from '../../../models/pool.model'
-import { drawCircle, drawText, getTooltipContainer } from '../../../utils/daily-chart/daily-chart.util'
-import { type PlotFunction } from '../../../models/plot-function.model'
-import { type PlotSelection } from '../../../models/plot-selection.model'
-import { type PlotOptions } from '../../../models/plot-options.model'
-import { type SuperpositionEvent } from '../../../models/superposition-event.model'
-import { SuperpositionEventSeverity } from '../../../models/enums/superposition-event-severity.enum'
-import { createIdGenerator } from '../../../utils/id-generator/id-generator.util'
-import { DailyPlotElement } from '../../../models/enums/daily-plot-element.enum'
-import { PLOT_DIMENSIONS } from '../../../models/constants/plot.constants'
+import { type Pool } from '../../../../models/pool.model'
+import { drawCircle, drawText, getTooltipContainer } from '../../../../utils/daily-chart/daily-chart.util'
+import { type PlotFunction } from '../../../../models/plot-function.model'
+import { type PlotSelection } from '../../../../models/plot-selection.model'
+import { type PlotOptions } from '../../../../models/plot-options.model'
+import { type SuperpositionEvent } from '../../../../models/superposition-event.model'
+import { SuperpositionEventSeverity } from '../../../../models/enums/superposition-event-severity.enum'
+import { createIdGenerator } from '../../../../utils/id-generator/id-generator.util'
+import { DailyPlotElement } from '../../../../models/enums/daily-plot-element.enum'
+import { PLOT_DIMENSIONS } from '../../../../models/constants/plot.constants'
 
 // ID generator for consistent element identification
 const idGen = createIdGenerator(DailyPlotElement.EventSuperposition)