+++ packages/yourloops/test/integration/mock/data.api.mock.ts
@@ -1716,7 +1716,7 @@ export const smallDataSet: Data = {
 
 export const noData: Data = {
   dataRange: ['', ''],
-  data: []
+  data: [] as unknown as MedicalData
 }
 
 export const smbgData: Data = {
@@ -1728,8 +1728,11 @@ export const smbgData: Data = {
     "cbg": [],
     "confidentialModes": [],
     "deviceParametersChanges": [],
+    "eatingShortlyEvents": [],
+    "iob": [],
     "messages": [],
     "meals": [],
+    "nightModes": [],
     "physicalActivities": [],
     "pumpSettings": [],
     "reservoirChanges": [],
@@ -2233,8 +2236,11 @@ export const timeInRangeStatsTrendViewData: Data = {
     ],
     "confidentialModes": [],
     "deviceParametersChanges": [],
+    "eatingShortlyEvents": [],
+    "iob": [],
     "messages": [],
     "meals": [],
+    "nightModes": [],
     "physicalActivities": [],
     "pumpSettings": [],
     "reservoirChanges": [],
@@ -2496,9 +2502,11 @@ export const twoWeeksOfCbg: Data = {
     ],
     "confidentialModes": [],
     "deviceParametersChanges": [],
+    "eatingShortlyEvents": [],
     "iob": [],
     "messages": [],
     "meals": [],
+    "nightModes": [],
     "physicalActivities": [],
     "pumpSettings": [],
     "reservoirChanges": [],